requirements.txt
streamlit==1.31.1
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
pyzbar==0.1.9
pandas==2.2.1